+++ app/build.gradle
@@ -57,6 +57,7 @@ dependencies {
     testImplementation 'junit:junit:4.13'
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "com.github.topjohnwu.libsu:core:2.5.1"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.9"
     // 脚本解析
     implementation 'org.apache-extras.beanshell:bsh:2.0b6'
     // androidx
