+++ app/src/main/java/me/kyuubiran/hook/BaseMultiConfigDelayableHook.java
@@ -16,7 +16,7 @@
  * along with this software.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package nil.nadph.qnotified.hook.kyuubiran;
+package me.kyuubiran.hook;
 
 import java.io.IOException;
