+++ app/src/main/java/me/kyuubiran/hook/RemoveQbossAD.java
@@ -1,18 +1,19 @@
-package nil.nadph.qnotified.hook.kyuubiran;
+package me.kyuubiran.hook;
 
 import android.os.Looper;
+import android.view.View;
 import android.widget.Toast;
 
 import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
 
 import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XposedBridge;
 import nil.nadph.qnotified.SyncUtils;
 import nil.nadph.qnotified.config.ConfigManager;
 import nil.nadph.qnotified.hook.BaseDelayableHook;
-import nil.nadph.qnotified.step.DexDeobfStep;
 import nil.nadph.qnotified.step.Step;
-import nil.nadph.qnotified.util.DexKit;
+import nil.nadph.qnotified.util.Initiator;
 import nil.nadph.qnotified.util.LicenseStatus;
 import nil.nadph.qnotified.util.Utils;
 
@@ -20,13 +21,13 @@ import static nil.nadph.qnotified.util.Utils.TOAST_TYPE_ERROR;
 import static nil.nadph.qnotified.util.Utils.getApplication;
 import static nil.nadph.qnotified.util.Utils.log;
 
-//屏蔽群聊界面一起嗨
-public class RemovePlayTogether extends BaseDelayableHook {
-    public static final String kr_remove_play_together = "kr_remove_play_together";
-    private static final RemovePlayTogether self = new RemovePlayTogether();
+//移除消息列表顶栏横幅广告
+public class RemoveQbossAD extends BaseDelayableHook {
+    public static final String kr_remove_qboss_ad = "kr_remove_qboss_ad";
+    private static final RemoveQbossAD self = new RemoveQbossAD();
     private boolean isInit = false;
 
-    public static RemovePlayTogether get() {
+    public static RemoveQbossAD get() {
         return self;
     }
 
@@ -44,14 +45,15 @@ public class RemovePlayTogether extends BaseDelayableHook {
     public boolean init() {
         if (isInit) return true;
         try {
-            for (Method m : DexKit.doFindClass(DexKit.C_TogetherControlHelper).getDeclaredMethods()) {
+            for (Method m : Initiator._QbossADImmersionBannerManager().getDeclaredMethods()) {
                 Class<?>[] argt = m.getParameterTypes();
-                if ("h".equals(m.getName()) && m.getReturnType() == void.class && argt.length == 0) {
+                if (m.getReturnType() == View.class && argt.length == 0 && !Modifier.isStatic(m.getModifiers())) {
                     XposedBridge.hookMethod(m, new XC_MethodHook() {
                         @Override
                         protected void beforeHookedMethod(MethodHookParam param) throws Throwable {
                             if (LicenseStatus.sDisableCommonHooks) return;
                             if (!isEnabled()) return;
+//                            Toast.makeText(getApplication(), "屏蔽消息列表横幅广告成功!", android.widget.Toast.LENGTH_SHORT).show();
                             param.setResult(null);
                         }
                     });
@@ -67,14 +69,14 @@ public class RemovePlayTogether extends BaseDelayableHook {
 
     @Override
     public Step[] getPreconditions() {
-        return new Step[]{new DexDeobfStep(DexKit.C_TogetherControlHelper)};
+        return new Step[0];
     }
 
     @Override
     public void setEnabled(boolean enabled) {
         try {
             ConfigManager mgr = ConfigManager.getDefaultConfig();
-            mgr.getAllConfig().put(kr_remove_play_together, enabled);
+            mgr.getAllConfig().put(kr_remove_qboss_ad, enabled);
             mgr.save();
         } catch (final Exception e) {
             Utils.log(e);
@@ -94,7 +96,7 @@ public class RemovePlayTogether extends BaseDelayableHook {
     @Override
     public boolean isEnabled() {
         try {
-            return ConfigManager.getDefaultConfig().getBooleanOrFalse(kr_remove_play_together);
+            return ConfigManager.getDefaultConfig().getBooleanOrFalse(kr_remove_qboss_ad);
         } catch (Exception e) {
             log(e);
             return false;
