+++ app/src/main/java/nil/nadph/qnotified/activity/SettingsActivity.java
@@ -37,7 +37,6 @@ import com.tencent.mobileqq.widget.BounceScrollView;
 import me.singleneuron.hook.ForceSystemCamera;
 import me.singleneuron.hook.NewRoundHead;
 import me.singleneuron.hook.adNoApplet;
-import me.singleneuron.util.QQVersion;
 import nil.nadph.qnotified.MainHook;
 import nil.nadph.qnotified.R;
 import nil.nadph.qnotified.config.ConfigItems;
@@ -45,7 +44,7 @@ import nil.nadph.qnotified.config.ConfigManager;
 import nil.nadph.qnotified.dialog.RepeaterIconSettingDialog;
 import nil.nadph.qnotified.dialog.RikkaDialog;
 import nil.nadph.qnotified.hook.*;
-import nil.nadph.qnotified.hook.kyuubiran.RemoveQbossAD;
+import me.kyuubiran.hook.*;
 import nil.nadph.qnotified.hook.rikka.RemoveMiniProgramAd;
 import nil.nadph.qnotified.ui.CustomDialog;
 import nil.nadph.qnotified.ui.HighContrastBorder;
