+++ app/src/main/java/nil/nadph/qnotified/hook/BaseDelayableHook.java
@@ -18,12 +18,13 @@
  */
 package nil.nadph.qnotified.hook;
 
+import me.kyuubiran.hook.*;
+import me.kyuubiran.hook.testhook.*;
 import me.singleneuron.hook.ForceSystemCamera;
 import me.singleneuron.hook.NewRoundHead;
 import me.singleneuron.hook.adNoApplet;
 import nil.nadph.qnotified.SyncUtils;
 import nil.nadph.qnotified.config.SwitchConfigItem;
-import nil.nadph.qnotified.hook.kyuubiran.*;
 import nil.nadph.qnotified.hook.rikka.*;
 import nil.nadph.qnotified.step.Step;
 import nil.nadph.qnotified.util.NonNull;
